requirements.txt
Flask
Flask-CORS
opencv-python
numpy
mysql-connector-python
deepface
